kbuild: fix optimization level choice default
The choice containing the CC_OPTIMIZE_FOR_PERFORMANCE symbol accidentally
added a "CONFIG_" prefix when trying to make it the default, selecting an
undefined symbol as the default.

The mistake is harmless here: Since the default symbol is not visible, the
choice falls back on using the visible symbol as the default instead,
which is CC_OPTIMIZE_FOR_PERFORMANCE, as intended.

A patch that makes Kconfig print a warning in this case has been submitted
separately: http://www.spinics.net/lists/linux-kbuild/msg15566.html

Link: http://lkml.kernel.org/r/1507074806-21577-1-git-send-email-ulfalizer@gmail.com
Signed-off-by: Ulf Magnusson <ulfalizer@gmail.com>
Acked-by: Arnd Bergmann <arnd@arndb.de>
Cc: Paul E . McKenney <paulmck@linux.vnet.ibm.com>
Cc: Nicolas Pitre <nicolas.pitre@linaro.org>
Cc: Tejun Heo <tj@kernel.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Kees Cook <keescook@chromium.org>
Cc: Daniel Mack <daniel@zonque.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>