kvm, mm: account kvm related kmem slabs to kmemcg
The kvm slabs can consume a significant amount of system memory and indeed
in our production environment we have observed that a lot of machines are
spending significant amount of memory that can not be left as system
memory overhead.  Also the allocations from these slabs can be triggered
directly by user space applications which has access to kvm and thus a
buggy application can leak such memory.  So, these caches should be
accounted to kmemcg.

Link: http://lkml.kernel.org/r/20171006010724.186563-1-shakeelb@google.com
Signed-off-by: Shakeel Butt <shakeelb@google.com>
Reviewed-by: Paolo Bonzini <pbonzini@redhat.com>
Cc: Anshuman Khandual <khandual@linux.vnet.ibm.com>
Cc: Radim Krčmář <rkrcmar@redhat.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: "H . Peter Anvin" <hpa@zytor.com>
Cc: Vladimir Davydov <vdavydov.dev@gmail.com>
Cc: Michal Hocko <mhocko@kernel.org>
Cc: Greg Thelen <gthelen@google.com>
Cc: Christian Borntraeger <borntraeger@de.ibm.com>
Cc: James Hogan <james.hogan@imgtec.com>
Cc: Christoffer Dall <christoffer.dall@linaro.org>
Cc: Paul Mackerras <paulus@ozlabs.org>
Cc: David Gibson <dgibson@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>