f2fs: simplify page iteration loops
In several places we want to iterate over all tagged pages in a mapping. 
However the code was apparently copied from places that iterate only over
a limited range and thus it checks for index <= end, optimizes the case
where we are coming close to range end which is all pointless when end ==
ULONG_MAX.  So just remove this dead code.

Link: http://lkml.kernel.org/r/20171009151359.31984-7-jack@suse.cz
Signed-off-by: Jan Kara <jack@suse.cz>
Reviewed-by: Daniel Jordan <daniel.m.jordan@oracle.com>
Reviewed-by: Chao Yu <yuchao0@huawei.com>
Cc: Jaegeuk Kim <jaegeuk@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>